refactor(Mileage, MileageTest, README): `주행거리`를 불변객체로 수정
- `주행거리`가 값으로만 사용되는 점을 고려하여 불변객체로 수정함
- 테스트 케이스 수정
- README.md 완료한 구현 기능 목록 체크

Mileage.java
MileageTest.java
README.md
