+++ README.md
@@ -1,15 +1,17 @@
 # java-racingcar-precourse
 
 ## 1. 구현할 기능 목록
-- [ ] `자동차`는 이름이 있다.
-- [ ] `자동차이름`은 최대 5자다.
-- [ ] `자동차`는 주행할 수 있다. (랜덤 라이브러리 사용)
-- [ ] `자동차`의 `주행 결과`는 `완주(FINISH), 미완주(RETIRE)` 중 하나가 된다.
+
+- [x] `자동차`는 이름이 있다.
+- [x] `자동차이름`은 최대 5자다.
+- [x] `자동차`는 주행할 수 있다. (랜덤 라이브러리 사용)
+- [x] `자동차`의 `주행 결과`는 `완주(FINISH), 미완주(RETIRE)` 중 하나가 된다.
 - [ ] `자동차경주`는 각 `랩(Lap)`으로 이루어진다.
 - [ ] 한 `랩(Lap)`을 진행하여 `자동차-랩-성적(CarLapRecord / 자동차, 주행결과, 주행거리)`을 기록할 수 있다.
 - [ ] 모든 `랩`이 끝난 `자동차경주`는, 마지막 `랩`의 `자동차-랩-성적`의 주행거리로 승자를 구분할 수 있다.
 - [ ] 형식에 맞게 입력한다. (자동차들의 이름, 자동차경주의 총 랩(Lap) 수)
 - [ ] 형식에 맞게 출력한다. (랩의 자동차 별 주행거리, 최종 우승자)
+- [x] `주행거리`는 `최대 랩 수` 이하로 존재한다.
 - 객체 간의 상호작용이 복잡해 보이므로, 최대한 작은 단위의 객체부터 구현과 테스트를 작성하자.
 
 ## 2. 주의사항
@@ -24,12 +26,12 @@
 의 형태로 기록한다.
 
 - 우승자는 반드시 한 명 이상이다.
-  - `우승자는 한 명 이상일 수 있다.`
+    - `우승자는 한 명 이상일 수 있다.`
 - 유효한 시도 횟수는 양의 정수와 0이다. 문자, 음수, 소수는 모두 예외처리 한다.
-  - `*몇 번*의 *이동*을 할 것인지를 입력`
-  - 이동, 즉 방향을 고려하지 않으며, 1 단위로 헤아릴 수 있는 물리량을 말한다.
-  - 0은 `0번의 이동`이 아닌 `이동하지 않음`이기 때문에 맥락상으로는 맞지 않는다. **하지만**, 규칙 상 이동 기회가 주어지고 나서 전혀 움직이지 않은 자동차들도 우승자가 되는 케이스가 존재한다. 즉 움직이지 않는 것도 대회가 정상적으로 수행된 걸로 판단하는 명백한 케이스가 있기 때문에, **0을 허용한다**. 
-
+    - `*몇 번*의 *이동*을 할 것인지를 입력`
+    - 이동, 즉 방향을 고려하지 않으며, 1 단위로 헤아릴 수 있는 물리량을 말한다.
+    - 0은 `0번의 이동`이 아닌 `이동하지 않음`이기 때문에 맥락상으로는 맞지 않는다. **하지만**, 규칙 상 이동 기회가 주어지고 나서 전혀 움직이지 않은 자동차들도 우승자가 되는 케이스가 존재한다. 즉
+      움직이지 않는 것도 대회가 정상적으로 수행된 걸로 판단하는 명백한 케이스가 있기 때문에, **0을 허용한다**.
 
 ## 애매한 동작의 규칙
 
@@ -46,31 +48,33 @@
 의 형태로 기록한다.
 
 - `""(empty string)` 이름을 허용할 것인가?
-  - 이름에 대한 제한은 `5자 이하`라는 것 뿐이다.
-  - 일반 사용자에게 있어서 공백은 이름이라고 보기 어렵다. 하지만 익명(이름을 제출하지 않은 선수)이라고 해석되기도 한다. 오롯이 선택의 문제다.
-  - 하지만 **출전자 입력이 `,`로 끝나는 경우**, 이것이 마무리 되지 않은(확실하게 출전자를 입력하지 않은) **구분자 종료 문자열인지 익명의 출전자가 마지막에 있는 것인지 불분명한 케이스**를 더욱 불러오게 된다.
-  - 따라서 **허용하지 않는다**.
+    - 이름에 대한 제한은 `5자 이하`라는 것 뿐이다.
+    - 일반 사용자에게 있어서 공백은 이름이라고 보기 어렵다. 하지만 익명(이름을 제출하지 않은 선수)이라고 해석되기도 한다. 오롯이 선택의 문제다.
+    - 하지만 **출전자 입력이 `,`로 끝나는 경우**, 이것이 마무리 되지 않은(확실하게 출전자를 입력하지 않은) **구분자 종료 문자열인지 익명의 출전자가 마지막에 있는 것인지 불분명한 케이스**를 더욱
+      불러오게 된다.
+    - 따라서 **허용하지 않는다**.
 - `" "(space)` 이름을 허용할 것인가?
-  - `" "` 이름에 대한 규정은 없으나, `"    "`나 `"     "` 등은 서로 식별되지 않는 이름이라 봐야 옳다.
-  - 또한 구분된 이름의 앞뒤는 trim하는 것이 맞다고 판단되는데(특히 앞), `abc, `는 마지막이 `""`가 입력된 케이스와 구분할 수 없다.
-  - 즉 **이름 사이가 공백인 경우만 허용하고**, 나머지는 `""`와 동일하게 취급하여 **허용하지 않는다**.
+    - `" "` 이름에 대한 규정은 없으나, `"    "`나 `"     "` 등은 서로 식별되지 않는 이름이라 봐야 옳다.
+    - 또한 구분된 이름의 앞뒤는 trim하는 것이 맞다고 판단되는데(특히 앞), `abc, `는 마지막이 `""`가 입력된 케이스와 구분할 수 없다.
+    - 즉 **이름 사이가 공백인 경우만 허용하고**, 나머지는 `""`와 동일하게 취급하여 **허용하지 않는다**.
 - 중복되는 이름을 허용할 것인가?
-  - 관련된 규칙이 전혀 없다.
-  - 실제로 동명이인은 얼마든지 있을 수 있다.
-  - 하지만 선수를 이름으로만 식별하고(유일한 식별자), 우승 시에 동명이인의 이름을 발표한다고 하면, 사실상 누가 우승했는지 알 수 없는 발표가 된다.
-  - 따라서 **이름의 중복을 허용하지 않는다**.
+    - 관련된 규칙이 전혀 없다.
+    - 실제로 동명이인은 얼마든지 있을 수 있다.
+    - 하지만 선수를 이름으로만 식별하고(유일한 식별자), 우승 시에 동명이인의 이름을 발표한다고 하면, 사실상 누가 우승했는지 알 수 없는 발표가 된다.
+    - 따라서 **이름의 중복을 허용하지 않는다**.
 - 영문 대소문자를 구분할 것인가?
-  - 구분하지 않을 이유가 없다(즉 **구분한다**). 승리자 표기 시에도 혼동이 없다고 판단된다.
-  - 다만 **동등 비교를 구현할 때 대소문자 비교에 유의하기로 한다**.
+    - 구분하지 않을 이유가 없다(즉 **구분한다**). 승리자 표기 시에도 혼동이 없다고 판단된다.
+    - 다만 **동등 비교를 구현할 때 대소문자 비교에 유의하기로 한다**.
 - 몇 대의 자동차를 출전 허용할 것인가? 몇 라운드의 랩을 주행할 것인가?
-  - 제한은 전혀 없다.
-  - 하지만 무제한으로 두었을 때 어플리케이션 실행에 지장을 줄 만큼 큰 성능 문제가 있을 거라고 예상된다. (O(n^m))
-  - 출전하는 **자동차는 5대**, **랩 수는 10랩**까지로 제한한다.
+    - 제한은 전혀 없다.
+    - 하지만 무제한으로 두었을 때 어플리케이션 실행에 지장을 줄 만큼 큰 성능 문제가 있을 거라고 예상된다. (O(n^m))
+    - 출전하는 **자동차는 5대**, **랩 수는 10랩**까지로 제한한다.
 - 최종 우승자가 여러 명일 경우 `,`로 구분할까, `, `로 구분할까?
-  - 문서의 설명에는 `,`로 구분하라 되어있고 예시에는 `, `로 나와있다.
-  - 하지만 예제 테스트 케이스를 보았을 때도 `,`로 되어있기 때문에, **우승자는 `,`로 구분한다**.
+    - 문서의 설명에는 `,`로 구분하라 되어있고 예시에는 `, `로 나와있다.
+    - 하지만 예제 테스트 케이스를 보았을 때도 `,`로 되어있기 때문에, **우승자는 `,`로 구분한다**.
 - 참가자가 한 명도 없는 케이스
-  - `우승자는 한 명 이상일 수 있다.` 문구에 의지하여 해석한다.
-  - 우승자는 한 명이나, 동등한 경우 한 명 이상일 수 있다.
-  - 한 명도 없는 경우가 불가능 하다고 명시되어 있지는 않으나, 우승 가능한 케이스를 설명하며 한 명(기본 케이스)외에 예외적인 케이스를 부연설명하고 있다. 다만 그 안에 '우승자가 없는 경우'가 포함되지 않았다. 우승자가 없는 유일한 케이스는 참가자가 한 명도 없는 경우 뿐이다.
-  - 즉 참가자가 한 명도 없는 케이스를 기대하지 않았다고 보는 편에 비중이 있다.
+    - `우승자는 한 명 이상일 수 있다.` 문구에 의지하여 해석한다.
+    - 우승자는 한 명이나, 동등한 경우 한 명 이상일 수 있다.
+    - 한 명도 없는 경우가 불가능 하다고 명시되어 있지는 않으나, 우승 가능한 케이스를 설명하며 한 명(기본 케이스)외에 예외적인 케이스를 부연설명하고 있다. 다만 그 안에 '우승자가 없는 경우'가 포함되지
+      않았다. 우승자가 없는 유일한 케이스는 참가자가 한 명도 없는 경우 뿐이다.
+    - 즉 참가자가 한 명도 없는 케이스를 기대하지 않았다고 보는 편에 비중이 있다.
